@@ -1,32 +1,28 @@
 {
   "name": "vuejs-demo-susantokun",
-  "version": "0.1.0",
+  "version": "2.0.0",
   "description": "A Vue.js project demo.susantokun.com",
-  "author": "susantokun <admin@susantokun.com>",
+  "author": "susantokun <susantokun.com@gmail.com>",
   "private": true,
+  "type": "module",
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
   },
   "dependencies": {
-    "axios": "^0.19.2",
-    "core-js": "^3.6.4",
-    "fomantic-ui": "^2.8.4",
-    "jquery-tablesort": "0.0.11",
-    "vue": "^2.6.11",
-    "vue-router": "^3.1.5",
-    "vuex": "^3.1.2"
+    "axios": "^1.8.3",
+    "fomantic-ui": "^2.9.4",
+    "fomantic-ui-css": "^2.9.4",
+    "jquery-tablesort": "^0.0.11",
+    "vue": "^3.5.13",
+    "vue-fomantic-ui": "^0.29.0",
+    "vue-router": "^4.5.0",
+    "vuex": "^4.1.0"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.2.0",
-    "@vue/cli-plugin-eslint": "~4.2.0",
-    "@vue/cli-plugin-router": "~4.2.0",
-    "@vue/cli-plugin-vuex": "~4.2.0",
-    "@vue/cli-service": "~4.2.0",
-    "babel-eslint": "^10.0.3",
-    "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^6.1.2",
-    "vue-template-compiler": "^2.6.11"
+    "@vitejs/plugin-vue": "^5.2.1",
+    "vite": "^6.2.1",
+    "vite-plugin-vue-devtools": "^7.7.2"
   }
 }
