@@ -1,34 +1,21 @@
-import Vue from "vue";
-import VueRouter from "vue-router";
-import Home from "../views/Home.vue";
-import NotFound from "../views/NotFound.vue";
-
-Vue.use(VueRouter);
+import About from '@/views/About.vue';
+import Home from '@/views/Home.vue';
+import NotFound from '@/views/NotFound.vue';
+import {
+  createMemoryHistory,
+  createRouter,
+  createWebHistory,
+} from 'vue-router';
 
 const routes = [
-  {
-    path: "/",
-    name: "Home",
-    component: Home
-  },
-  {
-    path: "/about",
-    name: "About",
-    // route level code-splitting
-    // this generates a separate chunk (about.[hash].js) for this route
-    // which is lazy-loaded when the route is visited.
-    component: () =>
-      import(/* webpackChunkName: "about" */ "../views/About.vue")
-  },
-  { path: "/404", component: NotFound },
-  { path: "*", redirect: "/404" }
-];
+  { path: '/', component: Home },
+  { path: '/about', component: About },
+  { path: '/:pathMatch(.*)*', name: 'NotFound', component: NotFound },
+]
 
-const router = new VueRouter({
-  mode: "history",
-  base: process.env.BASE_URL,
+const router = createRouter({
+  history: createWebHistory(),
   routes,
-  linkExactActiveClass: "active"
-});
+})
 
 export default router;
